bot/pyproject.toml
[tool.poetry]
name = "cafe-azu-bot-1"
version = "0.1.0"
description = "Телеграм бот для кафе АЗУ"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
asyncpg = "^0.28.0"
sqlalchemy = "1.4.50"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.7"}
python-telegram-bot-calendar = "^1.0.5"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pep8-naming = "^0.13.3"
flake8-broken-line = "^1.0.0"
flake8-return = "^1.2.0"
flake8-isort = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
